+++ test/rpc/faucet_client_test.dart
@@ -2,26 +2,12 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:sui/sui.dart';
 
 void main() {
-  test('request sui faucet v0', () async {
+  test('request sui faucet v2', () async {
     final faucet = FaucetClient(SuiUrls.faucetDev);
-    final resp = await faucet.requestSuiFromFaucetV0(
+    final resp = await faucet.requestSuiFromFaucetV2(
         '0x936accb491f0facaac668baaedcf4d0cfc6da1120b66f77fa6a43af718669973');
-    expect(resp.transferredGasObjects.isNotEmpty, true);
-  });
-
-  test('request sui faucet v1', () async {
-    final faucet = FaucetClient(SuiUrls.faucetDev);
-    final resp = await faucet.requestSuiFromFaucetV1(
-        '0x374114b640af9825ad85bd174f28ea38ed217c0740b4e5f588e8535dfbbe92e0');
-    expect(resp.task?.isNotEmpty, true);
-    final statusResp = await faucet.getFaucetRequestStatus(resp.task!);
-    expect(statusResp.status.status, BatchSendStatus.INPROGRESS);
-  });
-
-  test('get faucet request status', () async {
-    final faucet = FaucetClient(SuiUrls.faucetDev);
-    final resp = await faucet
-        .getFaucetRequestStatus("2442c39e-5cc5-4162-a7ef-e328cdc0332e");
-    expect(resp.status.status, BatchSendStatus.SUCCEEDED);
+    expect(resp.isSuccess, true);
+    expect(resp.coinsSent.isNotEmpty, true);
+    expect(resp.coinsSent.first.transferTxDigest.isNotEmpty, true);
   });
 }